Switch PhotoContext to SelectionContext
