+++ web/src/app/components/Photo/PhotoGridItem.tsx
@@ -8,7 +8,7 @@ import styles from './PhotoGridItem.module.css';
 
 // State
 import { getPhotoById } from '../../services/photo.service';
-import { usePhotoContext } from '../../contexts/PhotoContext';
+import { useSelectonContext } from '../../contexts/SelectionContext';
 import { useServiceContext } from '../../contexts/ServiceContext';
 
 const TIME_TO_DOWNLOAD = 1000;
@@ -20,7 +20,7 @@ function PhotoGridItem({ id, src }) {
     const originalSrc = React.useRef('');
     const timeoutId = React.useRef(null);
 
-    const { selectedThumbnails, setSelectedThumbnails, setLoadingLightBox } = usePhotoContext();
+    const { selectedThumbnails, setSelectedThumbnails, setLoadingLightBox } = useSelectonContext();
     const { useServer } = useServiceContext();
 
     React.useEffect(() => {
