+++ web/src/app/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { OverlayProvider } from './contexts/OverlayContext';
-import { PhotoProvider } from './contexts/PhotoContext';
+import { SelectionProvider } from './contexts/SelectionContext';
 import { ProgressProvider } from './contexts/ProgressContext';
 import { ServiceProvider } from './contexts/ServiceContext';
 import { ThemeProvider } from './contexts/ThemeContext';
@@ -15,11 +15,11 @@ function AppProvider() {
                 <ServiceProvider>
                     <ProgressProvider>
                         <OverlayProvider>
-                            <PhotoProvider>
+                            <SelectionProvider>
                                 {/** not ideal since photo states will bleed through all screen components under app 
                      But needed since the header needs to be aware of photo states | Possibly refactor later */}
                                 <App />
-                            </PhotoProvider>
+                            </SelectionProvider>
                         </OverlayProvider>
                     </ProgressProvider>
                 </ServiceProvider>
