+++ web/src/app/interfaces/contexts.interface.ts
@@ -1,10 +1,6 @@
-import { IThumbnail } from './photos.interface';
-
-export interface IPhotoContextValue {
+export interface ISelectionContextValue {
     selectedThumbnails: string[];
     setSelectedThumbnails: React.Dispatch<React.SetStateAction<string[]>>;
-    thumbnails: { [date: string]: { [photoId: string]: IThumbnail } };
-    setThumbnails: React.Dispatch<React.SetStateAction<{ [date: string]: { [photoId: string]: IThumbnail } }>>;
     loadingThumbnails: string[];
     setloadingThumbnails: React.Dispatch<React.SetStateAction<string[]>>;
     loadingLightBox: boolean;
